fix bindings and trace
