+++ bindings/src/stream_writer.rs
@@ -8,10 +8,10 @@
 // http://www.apache.org/licenses/LICENSE-2.0
 //
 
+use std::io::Error;
 cfg_if! {
     if #[cfg(feature = "python_binding")] {
-        use pravega_client::errors::SegmentWriterError;
-        use pravega_client::event_stream_writer::EventStreamWriter;
+        use pravega_client::event::writer::EventWriter;
         use pravega_client_shared::ScopedStream;
         use pravega_client::client_factory::ClientFactory;
         use pyo3::exceptions;
@@ -33,7 +33,7 @@ cfg_if! {
 #[pyclass]
 #[derive(new)]
 pub(crate) struct StreamWriter {
-    writer: EventStreamWriter,
+    writer: EventWriter,
     factory: ClientFactory,
     stream: ScopedStream,
 }
@@ -97,7 +97,7 @@ impl StreamWriter {
     #[args(event, routing_key = "None", "*")]
     pub fn write_event_bytes(&mut self, event: &[u8], routing_key: Option<&str>) -> PyResult<()> {
         // to_vec creates an owned copy of the python byte array object.
-        let write_future: tokio::sync::oneshot::Receiver<Result<(), SegmentWriterError>> = match routing_key {
+        let write_future: tokio::sync::oneshot::Receiver<Result<(), Error>> = match routing_key {
             Option::None => {
                 trace!("Writing a single event with no routing key");
                 self.factory
@@ -115,7 +115,7 @@ impl StreamWriter {
         let _guard = self.factory.get_runtime().enter();
         let timeout_fut = timeout(Duration::from_secs(TIMEOUT_IN_SECONDS), write_future);
 
-        let result: Result<Result<Result<(), SegmentWriterError>, RecvError>, _> =
+        let result: Result<Result<Result<(), Error>, RecvError>, _> =
             self.factory.get_runtime().block_on(timeout_fut);
         match result {
             Ok(t) => match t {
