+++ config/Cargo.toml
@@ -24,3 +24,5 @@ reqwest = { version = "0.11", features = ["json"]}
 base64 = "0.12"
 tokio = "1.1"
 async-trait = "0.1"
+tracing = "0.1"
+tracing-subscriber = "0.2"
