+++ config/src/lib.rs
@@ -35,6 +35,8 @@ use pravega_client_shared::PravegaNodeUri;
 use std::collections::HashMap;
 use std::env;
 use std::time::Duration;
+use tracing::{dispatcher, Dispatch, Level};
+use tracing_subscriber::FmtSubscriber;
 
 pub const MOCK_CONTROLLER_URI: (&str, u16) = ("localhost", 9090);
 const AUTH_METHOD: &str = "method";
@@ -48,6 +50,7 @@ const TLS_CERT_PATH: &str = "pravega_client_tls_cert_path";
 
 #[derive(Builder, Debug, Getters, CopyGetters, Clone)]
 #[builder(setter(into))]
+#[builder(build_fn(validate = "Self::validate"))]
 pub struct ClientConfig {
     #[get_copy = "pub"]
     #[builder(default = "u32::max_value()")]
@@ -98,6 +101,10 @@ pub struct ClientConfig {
     #[get_copy = "pub"]
     #[builder(default = "self.default_timeout()")]
     pub request_timeout: Duration,
+
+    #[get_copy = "pub"]
+    #[builder(default = "Level::INFO")]
+    pub log_level: Level,
 }
 
 impl ClientConfigBuilder {
@@ -141,6 +148,18 @@ impl ClientConfigBuilder {
     fn default_timeout(&self) -> Duration {
         Duration::from_secs(30)
     }
+
+    fn validate(&self) -> Result<(), String> {
+        let subscriber = FmtSubscriber::builder()
+            .with_ansi(true)
+            .with_max_level(Level::DEBUG)
+            .finish();
+
+        let my_dispatch = Dispatch::new(subscriber);
+        // this function can only be called once.
+        dispatcher::set_global_default(my_dispatch)
+            .map_err(|e| format!("failed to set tracing level globally: {:?}", e))
+    }
 }
 
 #[cfg(test)]
@@ -176,7 +195,7 @@ mod tests {
             .build()
             .unwrap();
 
-        assert_eq!(config.max_connections_in_pool(), u32::max_value() as u32);
+        assert_eq!(config.max_connections_in_pool(), u32::MAX as u32);
         assert_eq!(config.max_controller_connections(), 3u32);
         assert_eq!(config.connection_type(), ConnectionType::Tokio);
         assert_eq!(config.retry_policy(), RetryWithBackoff::default());
